- Changed slave chasing from using unit_walktobl to map_search_freecell + unit_walktoxy, since the previous behaviour makes all slaves always end up on the same cell. - Changed some function declarations to take x,y arguments as short rather than int.
git-svn-id: https://svn.code.sf.net/p/rathena/svn/trunk@6024 54d463be-8e91-2dee-dedb-b68131a5f0ec

+++ src/map/map.c
@@ -1380,7 +1380,7 @@ static int map_count_sub(struct block_list *bl,va_list ap)
  * &4 = there shouldn't be any players around the target tile (use the no_spawn_on_player setting)
  *------------------------------------------
  */
-int map_search_freecell(struct block_list *src, int m, int *x,int *y, int rx, int ry, int flag) {
+int map_search_freecell(struct block_list *src, int m, short *x,short *y, int rx, int ry, int flag) {
 	int tries, spawn=0;
 	int bx, by;
 	int rx2 = 2*rx+1;
